♻️ Refactor: 재사용 위한 userSocketMap 리팩토링
